Separated header files to include directory
