+++ src/CMakeLists.txt
@@ -1,0 +1,7 @@
+# add library
+add_library(libkilo++
+  Editor.cpp
+  EditorUtils.cpp)
+
+# add include directories
+target_include_directories(libkilo++ PUBLIC ${CMAKE_SOURCE_DIR}/include)
