+++ src/EditorUtils.cpp
@@ -1,4 +1,4 @@
-#include "EditorUtils.hpp"
+#include "kilo++/EditorUtils.hpp"
 
 #include <cerrno>
 #include <cstdio>
